requirements.txt
a-pytube-fork-for-spotdl-users==9.6.4
aiofiles==0.6.0
antispaminc==3.8
aiohttp==3.7.3
apiclient==1.0.4
APScheduler==3.6.3
aria2p==0.9.1
asgiref==3.3.1
async==0.6.2
async-timeout==3.0.1
asyncurban==0.3.3
attrs==20.3.0
beautifulsoup4==4.9.3
blob==0.16
borg==2012.4.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
cfscrape==2.1.1
chardet==3.0.4
click==7.1.2
coffeehouse==2.2.2
colorama==0.4.4
Command==0.1.0
covid==2.4.0
cowpy==1.1.0
cryptography==3.2.1
cryptocompare==0.7.3
Cython==0.29.21
decorator==4.4.2
Deprecated==1.2.10
Django==3.1.3
dnspython==2.0.0
emoji==0.6.0
feedparser==6.0.2
faker==5.0.0
gcg==0.2.0
geographiclib==1.50
geopy==2.0.0
gitdb==4.0.5
githubpy==1.1.0
gitpy==0.0.5
gitpython==3.1.11
glitch-this==1.0.1
google==3.0.0
google-api-core==1.23.0
google-api-python-client==1.8.0
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-images-download==2.8.0
googleapis-common-protos==1.52.0
googletrans==4.0.0-rc1
gTTS==2.2.1
h11==0.9.0
h2==3.2.0
hachoir==3.1.1
heroku3==4.2.3
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
humanize==3.1.0
hyperframe==5.2.0
idna==2.10
image==1.5.33
iplookup==1.0.5
Jinja2==2.11.2
JustWatch==0.5.1
libmagic
loguru==0.5.3
lottie==0.6.6
lxml==4.5.2
lyricsgenius==2.0.2
MarkupSafe==1.1.1
multidict==5.0.2
mutagen==1.45.1
mpmath==1.1.0
numpy==1.19.4
oauth2client==4.1.3
opencv-contrib-python==4.4.0.46
password-strength==0.0.3.post2
Pillow==6.2.2
plac==1.2.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.0.2
pycparser==2.20
pydantic==1.7.2
pyfiglet==0.8.post1
PyGithub==1.53
pyjokes==0.6.0
PyJWT==1.7.1
pylast==4.0.0
PyLyrics==1.1.0
pyshorteners==1.0.1
pySmartDL==1.3.4
pytesseract==0.3.6
python-barcode==0.13.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-git==2018.2.1
python-github==0.1.0.dev0
python-magic==0.4.18
python-telegram-bot==13.0
pytube==10.0.0
pytz==2020.4
qrcode==6.1
rapidfuzz==0.13.3
redis==3.5.3
regex==2020.11.13
requests==2.25.0
rfc3986==1.4.0
rsa==4.6
scipy==1.5.4
search-engine-parser==0.4.2
selenium==3.5.0
semver==2.13.0
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
smmap==3.0.4
sniffio==1.2.0
soupsieve==2.0.1
spamwatch==0.3.0
speedtest==0.0.1
spotdl==3.1.4
spotify-token==1.0.0
spotipy==2.16.1
sqlalchemy==1.2
sqlparse==0.4.1
telegraph==1.4.1
telethon==1.17
tornado==6.1
tqdm==4.53.0
tswift==0.7.0
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urbandict==0.6.1
uritemplate==3.0.1
urllib3==1.26.2
uwsgidecorators==1.1.0
validators==0.18.1
websocket-client==0.57.0
webdriver-manager
wget==3.2
wikipedia==1.4.0
win32-setctime==1.0.3
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2020.11.24
youtube-search==1.1.1
youtube-search-python==1.3.2
pymediainfo==5.0.3
deep_translator==1.3.2
pokedex.py==1.1.2
cryptosteganography==0.8.3
yahoo_fin==0.8.6
langdetect==1.0.8
pybase64==1.0.2
tinydb==3.2.1
pyTelegramBotAPI==2.1.5
pygments==2.7.1
pandas==1.2.1
